fix: point 구조 수정 및 validator 추가
