+++ src/main/java/io/hhplus/tdd/database/PointHistoryTable.java
@@ -2,7 +2,7 @@ package io.hhplus.tdd.database;
 
 
 import io.hhplus.tdd.point.PointHistory;
-import io.hhplus.tdd.point.TransactionType;
+import io.hhplus.tdd.point.dto.TransactionType;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
