+++ src/main/java/io/hhplus/tdd/point/PointHistory.java
@@ -1,5 +1,7 @@
 package io.hhplus.tdd.point;
 
+import io.hhplus.tdd.point.dto.TransactionType;
+
 public record PointHistory(
         long id,
         long userId,
