+++ src/main/java/io/hhplus/tdd/point/controller/PointController.java
@@ -1,5 +1,8 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.controller;
 
+import io.hhplus.tdd.point.PointHistory;
+import io.hhplus.tdd.point.service.PointService;
+import io.hhplus.tdd.point.dto.UserPoint;
 import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -38,6 +41,6 @@ public class PointController {
     @PatchMapping("{id}/use")
     public UserPoint use(@PathVariable long id, @RequestBody long amount) {
         log.info("포인트 사용: id={} amount={}", id, amount);
-        return pointService.userPoint(id, amount);
+        return pointService.usePoint(id, amount);
     }
 }
